requirements.txt
paramiko==3.4.0
requests==2.31.0
python-dotenv==1.0.1
tqdm